+++ src/ReadInputData.h
@@ -4,45 +4,12 @@
 #define READINPUTDATA_H
 
 #include "Core_Export.h"
-#include <vector>
-#include <stdexcept>
-#include <cctype> // 用于isspace()
-#include <array>
-#include <iostream>
+#include "BoundaryData.h"
 #include <filesystem>
+#include <string>
+#include <fstream>
 
 namespace Poisson{
-
-/// @brief 表示单条边界信息的结构体
-struct SingleEdge {
-    std::string value = "";  // 边界函数的表达式
-    std::array<double, 2> range = {0.0, 1.0}; // 边界作用范围【min,max】
-
-    SingleEdge() = default;
-
-    SingleEdge(const std::string& val, const std::array<double, 2>& rng)
-        : value(val), range(rng){}
-};
-
-/// @brief 长方形区域中各种边界条件集合的结构体
-struct BoundaryConditionData {
-    SingleEdge AB;  // 底边 (x轴方向)
-    SingleEdge AD;  // 左边 (y轴方向)
-    SingleEdge BC;  // 右边 (y轴方向)
-    SingleEdge CD;  // 顶边 (x轴方向)
-
-    BoundaryConditionData() = default;
-
-    BoundaryConditionData(const std::string& ab, const std::array<double, 2>& ab_range,
-                         const std::string& ad, const std::array<double, 2>& ad_range,
-                         const std::string& bc, const std::array<double, 2>& bc_range,
-                         const std::string& cd, const std::array<double, 2>& cd_range)
-        : AB(ab, ab_range), 
-          AD(ad, ad_range), 
-          BC(bc, bc_range), 
-          CD(cd, cd_range) {}
-};
-
 
 /// @brief 处理输入文件，读取计算参数
 struct ReadInputData {
@@ -55,7 +22,7 @@ struct ReadInputData {
     std::string mesh_type = "triangle";///< 网格类型
 
     // 边界条件
-    BoundaryConditionData bc;             ///< 边界条件
+    BoundaryConditionInfo bc;             ///< 边界条件
 
     // 函数定义
     std::string initial_guess = "1.0"; ///< 初始猜测
