+++ src/RunPoissonSolver.h
@@ -4,12 +4,8 @@
 #define RUN_POISSON_SOLVER_H
 
 #include "Core_Export.h"
-#include "Help.h"
-#include "Timer.h"
-#include "Mesh.h"
-#include "BoundaryCondition.h"
-#include "PoissonSolver.h"
 #include "ReadInputData.h"
+#include "Timer.h"
 #include <filesystem>
 #include <memory>
 
@@ -29,9 +25,6 @@ public:
 private:
     std::filesystem::path m_jsonPath;
     std::unique_ptr<ReadInputData> m_readInputData;
-    std::unique_ptr<Mesh> m_mesh;
-    std::unique_ptr<BoundaryCondition> m_boundaryCondition;
-    std::unique_ptr<PoissonSolver> m_solver;
     Timer m_timer;
 };
